src/cloud_functions/result_processor/requirements.txt
      
# Requirements for the Result Processor function
google-cloud-storage
Pillow
pandas

    
